third

+++ ExpressUrl/drizzle/0000_dark_punisher.sql
@@ -42,7 +42,7 @@ CREATE TABLE `ShortLinks` (
 );
 --> statement-breakpoint
 CREATE TABLE `registerusers` (
-	`id` serial AUTO_INCREMENT NOT NULL,
+	`id` bigint AUTO_INCREMENT NOT NULL,
 	`name` varchar(255) NOT NULL,
 	`email` varchar(255) NOT NULL,
 	`password` varchar(255),
